refactor: 타입힌트 오류수정
정수가 아닌 실수이므로 타입힌트를 int에서 float로 변경

+++ moview/utils/data_manager.py
@@ -71,7 +71,7 @@ class ChatManager:
         callback_handler: Type[BaseCallbackHandler] = CustomCallbackHandler(),
         streaming: bool = True,
         model: str = "gpt-3.5-turbo",
-        temperature: int = 0.5
+        temperature: float = 0.5
     ):
         """
